-expa new counting method

+++ 100GreatestThings/Common/CommonUserDefaults.h
@@ -18,7 +18,7 @@
 
 @property (nonatomic) BOOL flagNotFirstLaunch;
 @property (nonatomic) int level;
-@property (nonatomic) int expa;
+@property (nonatomic) float expa;
 @property (nonatomic) float maxenergy;
 @property (nonatomic) float energyrange;
 @property (nonatomic,strong) NSDate* lastlaunchdate;
